requirements.txt
jinja2
pyyaml
requests
kubernetes